Move primitive gates header into subdirectory
