+++ include/mini-qiskit/circuit/circuit.hpp
@@ -14,7 +14,7 @@
 #include "mini-qiskit/common/matrix2x2.hpp"
 #include "mini-qiskit/common/utils.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
-#include "mini-qiskit/primitive_gate.hpp"
+#include "mini-qiskit/gates/primitive_gate.hpp"
 #include "mini-qiskit/circuit/control_flow.hpp"
 
 namespace mqis
