+++ include/mini-qiskit/circuit_operations/make_controlled_circuit.hpp
@@ -8,7 +8,7 @@
 #include "mini-qiskit/common/utils.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
 #include "mini-qiskit/gates/multiplicity_controlled_u_gate.hpp"
-#include "mini-qiskit/primitive_gate.hpp"
+#include "mini-qiskit/gates/primitive_gate.hpp"
 
 
 namespace impl_mqis
