+++ include/mini-qiskit/common/print.hpp
@@ -10,7 +10,7 @@
 #include <tuple>
 #include <unordered_map>
 
-#include "mini-qiskit/primitive_gate.hpp"
+#include "mini-qiskit/gates/primitive_gate.hpp"
 #include "mini-qiskit/common/matrix2x2.hpp"
 #include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/state/state.hpp"
